+++ src/advanced_graph_algorithms/eulerian_path.c
@@ -1,4 +1,4 @@
-#include "graph_traversals.h"
+#include "advanced_graph_algorithms.h"
 #include <stdbool.h>
 #include <stdlib.h>
 #include <string.h>
