+++ tests/advanced_graph_algorithms/test_dinic.c
@@ -1,13 +1,13 @@
-#include "../../src/graph_traversals/edmonds_karp.c"
-#include "graph_traversals.h"
+#include "../../src/advanced_graph_algorithms/dinic.c"
+#include "advanced_graph_algorithms.h"
 #include <assert.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
-void test_ek_simple()
+void test_dinic_simple()
 {
-    printf("Running test_ek_simple...\n");
+    printf("Running test_dinic_simple...\n");
     weightedGraph* g = create_weightedGraph(4);
     add_edge_directed(g, 0, 1, 3);
     add_edge_directed(g, 0, 2, 2);
@@ -15,16 +15,16 @@ void test_ek_simple()
     add_edge_directed(g, 2, 3, 3);
     add_edge_directed(g, 1, 2, 1);
 
-    int flow = edmonds_karp(g, 0, 3);
+    int flow = dinic(g, 0, 3);
     assert(flow == 5);
 
     free_weightedGraph(g);
-    printf("test_ek_simple passed.\n");
+    printf("test_dinic_simple passed.\n");
 }
 
 int main()
 {
-    test_ek_simple();
-    printf("All Edmonds-Karp Tests Passed!\n");
+    test_dinic_simple();
+    printf("All Dinic Tests Passed!\n");
     return 0;
 }
